@@ -1,16 +1,15 @@
 ﻿using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 
-namespace BookModels
-{
-    public record Book(
-        [property: Required]
-        [property: MaxLength(50)]
-        [property: DisplayName("Title")]
-        string Title,
-        [property: Required]
-        [property:MaxLength(50)]
-        [property:DisplayName("Publisher")]
-        string Publisher,
-        int BookId = 0);
-}
+namespace BookModels;
+
+public record Book(
+    [property: Required]
+    [property: MaxLength(50)]
+    [property: DisplayName("Title")]
+    string Title,
+    [property: Required]
+    [property:MaxLength(50)]
+    [property:DisplayName("Publisher")]
+    string Publisher,
+    int BookId = 0);
