@@ -1,38 +1,33 @@
 ﻿using BookModels;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using System.Threading.Tasks;
 
-namespace BooksViews
+namespace BooksViews;
+
+public class CreateModel : PageModel
 {
-    public class CreateModel : PageModel
+    private readonly BooksContext _context;
+
+    public CreateModel(BooksContext context) => _context = context;
+
+    public IActionResult OnGet()
     {
-        private readonly BookModels.BooksContext _context;
+        return Page();
+    }
 
-        public CreateModel(BookModels.BooksContext context)
-        {
-            _context = context;
-        }
+    [BindProperty]
+    public Book? Book { get; set; }
 
-        public IActionResult OnGet()
+    // To protect from overposting attacks, see https://aka.ms/RazorPagesCRUD
+    public async Task<IActionResult> OnPostAsync()
+    {
+        if (!ModelState.IsValid || Book is null)
         {
             return Page();
         }
+        _context.Books.Add(Book);
+        await _context.SaveChangesAsync();
 
-        [BindProperty]
-        public Book? Book { get; set; }
-
-        // To protect from overposting attacks, see https://aka.ms/RazorPagesCRUD
-        public async Task<IActionResult> OnPostAsync()
-        {
-            if (!ModelState.IsValid || Book is null)
-            {
-                return Page();
-            }
-            _context.Books.Add(Book);
-            await _context.SaveChangesAsync();
-
-            return RedirectToPage("./Index");
-        }
+        return RedirectToPage("./Index");
     }
 }
