@@ -1,76 +1,67 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using BookModels;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
-using BookModels;
 
-namespace BooksViews
+namespace BooksViews;
+
+public class EditModel : PageModel
 {
-    public class EditModel : PageModel
-    {
-        private readonly BookModels.BooksContext _context;
+    private readonly BooksContext _context;
 
-        public EditModel(BookModels.BooksContext context)
+    public EditModel(BooksContext context) => _context = context;
+
+    [BindProperty]
+    public Book? Book { get; set; }
+
+    public async Task<IActionResult> OnGetAsync(int? id)
+    {
+        if (id == null)
         {
-            _context = context;
+            return NotFound();
         }
 
-        [BindProperty]
-        public Book? Book { get; set; }
+        Book = await _context.Books.FirstOrDefaultAsync(m => m.BookId == id);
 
-        public async Task<IActionResult> OnGetAsync(int? id)
+        if (Book == null)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            return NotFound();
+        }
+        return Page();
+    }
 
-            Book = await _context.Books.FirstOrDefaultAsync(m => m.BookId == id);
-
-            if (Book == null)
-            {
-                return NotFound();
-            }
+    // To protect from overposting attacks, enable the specific properties you want to bind to.
+    // For more details, see https://aka.ms/RazorPagesCRUD.
+    public async Task<IActionResult> OnPostAsync()
+    {
+        if (!ModelState.IsValid)
+        {
             return Page();
         }
 
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see https://aka.ms/RazorPagesCRUD.
-        public async Task<IActionResult> OnPostAsync()
+        _context.Attach(Book).State = EntityState.Modified;
+
+        try
         {
-            if (!ModelState.IsValid)
+            await _context.SaveChangesAsync();
+        }
+        catch (DbUpdateConcurrencyException)
+        {
+            if (Book is null || !BookExists(Book.BookId))
             {
-                return Page();
+                return NotFound();
             }
-
-            _context.Attach(Book).State = EntityState.Modified;
-
-            try
+            else
             {
-                await _context.SaveChangesAsync();
+                throw;
             }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (Book is null || !BookExists(Book.BookId))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
-
-            return RedirectToPage("./Index");
         }
 
-        private bool BookExists(int id)
-        {
-            return _context.Books.Any(e => e.BookId == id);
-        }
+        return RedirectToPage("./Index");
+    }
+
+    private bool BookExists(int id)
+    {
+        return _context.Books.Any(e => e.BookId == id);
     }
 }
