@@ -10,7 +10,7 @@
         </tr>
     </thead>
     <tbody>
-        @foreach (var ev in Model)
+        @foreach (var ev in Model!)
         {
             <tr>
                 <td>@ev.Date.ToString("d")</td>
