@@ -1,6 +1,3 @@
-﻿using System;
+﻿namespace WebAppSample.Models;
 
-namespace WebAppSample.Models
-{
-    public record MenuItem(int Id, string Text, double Price, DateTime Date);
-}
+public record MenuItem(int Id, string Text, double Price, DateTime Date);
