@@ -29,6 +29,21 @@
                             <a class="nav-link text-dark" asp-area="" asp-page="/Index">Home</a>
                         </li>
                         <li class="nav-item">
+                            <a class="nav-link text-dark" asp-area="" asp-page="/ShowBook" asp-route-title="Professional C#" asp-route-publisher="Wrox">Book</a>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link text-dark" asp-area="" asp-page="/UseMarkdown">Markdown</a>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link text-dark" asp-area="" asp-page="UseMarkdownAttribute">Markdown 2</a>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link text-dark" asp-area="" asp-page="UseTableTagHelper">Tag Helper</a>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link text-dark" asp-area="" asp-page="UseViewComponent">View Component</a>
+                        </li>
+                        <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-page="/Privacy">Privacy</a>
                         </li>
                     </ul>
@@ -44,7 +59,7 @@
 
     <footer class="border-top footer text-muted">
         <div class="container">
-            &copy; 2021 - WebAppSample - <a asp-area="" asp-page="/Privacy">Privacy</a>
+            &copy; 2022 - WebAppSample - <a asp-area="" asp-page="/Privacy">Privacy</a>
         </div>
     </footer>
 
